+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Latest
 * Improved KOTH scoreboard and added disappearing smoke workaround
 * Added steersman zig-zag enable/disable function
+* Added "Hold Position" / "Resume Sailing" options to manually controlled steersman groups
 ## 1.8a
 * Created common comms management logic in HeLMS
 * Steersman comms control option added